pyproject.toml
[project]
name = "machwave-api"
version = "0.1.0"
description = "REST API for the machwave internal ballistics platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "firebase-admin>=6.5.0",
    "google-cloud-storage>=2.17.0",
    "google-cloud-run>=0.10.13",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "machwave==0.1.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
    "S101",  # allow assert in tests
    "B008",  # Depends() in argument defaults is idiomatic FastAPI
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
